Rename header files for testpaint
