+++ test/testpaint/TestPaint.hpp
@@ -28,7 +28,7 @@ extern "C"
     #include <openrct2/paint/tile_element/TileElement.h>
 }
 
-#include "addresses.h"
+#include "Addresses.h"
 
 #define gRideEntries                RCT2_ADDRESS(0x009ACFA4, rct_ride_entry*)
 #define gSupportSegments            RCT2_ADDRESS(0x0141E9B4, support_height)
