+++ test/testpaint/generate.cpp
@@ -26,7 +26,7 @@
 #include "String.hpp"
 #include "Utils.hpp"
 
-#include "data.h"
+#include "Data.h"
 #include <openrct2/interface/Viewport.h>
 #include <openrct2/rct2/RCT2.h>
 #include <openrct2/ride/Ride.h>
